Lot Of Updates in This Commit
1)Added Constructor That can help in assigning foreign key to other
table
2)Edited Masters and add demand,project and candidate master(May be
imoorved in next update)
3)Understand how to get data by passing object in DAO
4)Super Constructor Update
5)Single Employee Multiple Allowance Proper Flow Completed

+++ src/main/java/com/HRMS/services/impl/EmployeeAllowanceImpl.java
@@ -9,7 +9,9 @@ import org.springframework.stereotype.Service;
 
 import com.HRMS.controller.EmployeeAllowanceController;
 import com.HRMS.dao.EmployeeAllowanceDAO;
+import com.HRMS.model.AllowanceMaster;
 import com.HRMS.model.EmployeeAllowance;
+import com.HRMS.model.EmployeeMaster;
 import com.HRMS.services.EmployeeAllowanceService;
 
 @Service
@@ -20,16 +22,17 @@ public class EmployeeAllowanceImpl implements EmployeeAllowanceService {
 	@Autowired
 	private EmployeeAllowanceDAO EADAO;
 
-	public List<EmployeeAllowance> EmployeeAllAllowance(int empid) {
+	public List<EmployeeAllowance> EmployeeAllAllowance(int employeeId) {
 		try {
 
-			List<EmployeeAllowance> findByEmpid_empid = EADAO.findByEmployeeId(empid);
-			return findByEmpid_empid;
+			EmployeeMaster employeeMaster = new EmployeeMaster();
+			employeeMaster.setEmpId(employeeId);
+			return EADAO.findByEmployeeId(employeeMaster);
 
 		} catch (Exception e) {
 			e.printStackTrace();
-			logger.error(e);
-			return null;
+			logger.error("Error For Fetching Allowance for EmpId" + employeeId, e);
+			throw new RuntimeException("Fail to get EmployeeAllowance" + e);
 		}
 
 	}
@@ -37,40 +40,78 @@ public class EmployeeAllowanceImpl implements EmployeeAllowanceService {
 	@Override
 	public EmployeeAllowance assignallowance(EmployeeAllowance empallowance) {
 		try {
-			EmployeeAllowance employeeAllowance = this.EADAO.save(empallowance);
-			return employeeAllowance;
+			return this.EADAO.save(empallowance);
 		} catch (Exception e) {
 			e.printStackTrace();
-			logger.error(e);
+			logger.error("Error For Assigning Allowance for Empid" + empallowance.getEmployeeId().getEmpId(), e);
+			throw new RuntimeException("Failed to assign allowance" + e);
+		}
+
+	}
+
+	@Override
+	public EmployeeAllowance updateallowance(EmployeeAllowance empallowance) {
+		try {
+
+			EmployeeMaster empId = empallowance.getEmployeeId();
+			AllowanceMaster Alloid = empallowance.getAllowanceid();
+
+			EmployeeAllowance employeeAllowance = this.EADAO.findByEmployeeIdAndAllowanceid(empId, Alloid);
+
+			if (employeeAllowance == null) {
+				return null;
+			}
+
+			int ID = employeeAllowance.getId();
+			empallowance.setId(ID);
+			return this.EADAO.save(empallowance);
+
+		} catch (Exception e) {
+			e.printStackTrace();
+			logger.error("Error Updating Allowance for EmployeeId" + empallowance.getEmployeeId().getEmpId());
 			return null;
 		}
 
 	}
 
 	@Override
-	public EmployeeAllowance updateallowance(int empid, EmployeeAllowance empallowance, int allowance_id) {
+	public boolean deleteallowance(EmployeeAllowance empallowance) {
 		try {
-			empallowance.setEmployeeId(empid);
-			empallowance.setAllowanceId(allowance_id);
-			EmployeeAllowance employeeAllowance = this.EADAO.save(empallowance);
-			return employeeAllowance;
-		} catch (Exception e) {
-			e.printStackTrace();
-			logger.error(e);
-			return null;
-		}
+			EmployeeMaster empId = empallowance.getEmployeeId();
+			AllowanceMaster Alloid = empallowance.getAllowanceid();
 
-	}
+			EmployeeAllowance employeeAllowance = this.EADAO.findByEmployeeIdAndAllowanceid(empId, Alloid);
 
-	@Override
-	public void deleteallowance(EmployeeAllowance empallowance) {
-		try {
+			if (employeeAllowance == null) {
+				return false;
+			}
+			
+			int ID = employeeAllowance.getId();
+			empallowance.setId(ID);
 			this.EADAO.delete(empallowance);
+			return true;
+			
 		} catch (Exception e) {
 			e.printStackTrace();
 			logger.error(e);
+			return false;
 		}
 
+	}
+
+	@Override
+	public boolean CheckEmployee(EmployeeAllowance empallowance) {
+
+		EmployeeMaster empId = empallowance.getEmployeeId();
+		AllowanceMaster Alloid = empallowance.getAllowanceid();
+
+		EmployeeAllowance employeeAllowance = this.EADAO.findByEmployeeIdAndAllowanceid(empId, Alloid);
+
+		if (employeeAllowance == null) {
+			return false;
+		}
+		
+		return true;
 	}
 
 }
